+++ src/editor/component-editor/SegmentedControlEditor.svelte
@@ -207,7 +207,7 @@
     {component}
   >
     {#snippet extraPropertyRows(stateName)}
-      {#if stateName === 'hover'}
+      {#if stateName === 'hover option'}
         <div class="property-row" data-token-variables="--segmentedcontrol-hover-tint-enabled">
           <span class="property-label">tint layer</span>
           <Toggle checked={tintOn} onchange={setTintOn} />