+++ src/system/components/SideNavigation.svelte
@@ -265,7 +265,7 @@
     --sidenavigation-panel-item-padding: var(--space-32);
     --sidenavigation-panel-footer-gap: var(--space-16);
     --sidenavigation-panel-open-width: calc(var(--space-64) * 4);
-    --sidenavigation-panel-closed-width: var(--space-48);
+    --sidenavigation-panel-closed-width: calc(var(--space-64) * 0.75);
     --sidenavigation-open-duration: var(--duration-200);
     --sidenavigation-open-easing: var(--ease-out-quart);
     --sidenavigation-close-duration: var(--duration-150);